requirements.txt

Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-CORS==4.0.0
Flask-Limiter==3.5.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
google-generativeai==0.8.3
PyJWT==2.8.0
requests==2.31.0
beautifulsoup4==4.12.3
Pillow==10.2.0
numpy==1.26.4
gunicorn==21.2.0
werkzeug==3.0.1
SQLAlchemy==2.0.23
